Changes to isolate lgc/util
Part of making the LGC subdirectories more cleanly modularized.

* Moved PipelineShaders from util to state
* Moved ShaderStage-dependent funcs out of Internal.h and into a new
  state/ShaderStage.cpp and .h, also commoning up multiple occurrences
  of code to set shader stage metadata on a new function
* Moved lgcNames from util/Internal.h to state/Defs.h
* Also tidied up #includes in Internal.h
