+++ lgc/patch/PatchCopyShader.cpp
@@ -250,10 +250,8 @@ bool PatchCopyShader::runOnModule(Module &module) {
     builder.CreateBr(endBlock);
   }
 
-  // Add execution model metadata to the function.
-  auto execModelMeta = ConstantAsMetadata::get(ConstantInt::get(Type::getInt32Ty(*m_context), ShaderStageCopyShader));
-  auto execModelMetaNode = MDNode::get(*m_context, execModelMeta);
-  entryPoint->addMetadata(lgcName::ShaderStageMetadata, *execModelMetaNode);
+  // Set the shader stage on the new function.
+  setShaderStage(entryPoint, ShaderStageCopyShader);
 
   // Tell pipeline state there is a copy shader.
   m_pipelineState->setShaderStageMask(m_pipelineState->getShaderStageMask() | (1U << ShaderStageCopyShader));
